+++ mkahawa-srv/src/SettingsBox.cpp
@@ -453,7 +453,7 @@ SettingsBox::onSetTarif(FXObject*,FXSelector,void*)
     FXFoldingItem *sitem = tlist->getCurrentItem();
 
     if (sitem){
-      stariff = (int)(sitem->getData());
+      stariff = (long)(sitem->getData());
       char *tName = CCL_tarif_name_get(stariff);
       lblTariff->setText(tName);
       CCL_free(tName);
